Test cases and clean up
