+++ codefresh/provider.go
@@ -38,18 +38,17 @@ func Provider() *schema.Provider {
 			"codefresh_users":               dataSourceUsers(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
-			"codefresh_account":             resourceAccount(),
-			"codefresh_account_admins":      resourceAccountAdmins(),
-			"codefresh_api_key":             resourceApiKey(),
-			"codefresh_context":             resourceContext(),
-			"codefresh_idp_accounts":        resourceIDPAccounts(),
-			"codefresh_permission":          resourcePermission(),
-			"codefresh_pipeline":            resourcePipeline(),
-			"codefresh_project":             resourceProject(),
-			"codefresh_step_types":          resourceStepTypes(),
-			"codefresh_step_types_versions": resourceStepTypesVersions(),
-			"codefresh_user":                resourceUser(),
-			"codefresh_team":                resourceTeam(),
+			"codefresh_account":        resourceAccount(),
+			"codefresh_account_admins": resourceAccountAdmins(),
+			"codefresh_api_key":        resourceApiKey(),
+			"codefresh_context":        resourceContext(),
+			"codefresh_idp_accounts":   resourceIDPAccounts(),
+			"codefresh_permission":     resourcePermission(),
+			"codefresh_pipeline":       resourcePipeline(),
+			"codefresh_project":        resourceProject(),
+			"codefresh_step_types":     resourceStepTypes(),
+			"codefresh_user":           resourceUser(),
+			"codefresh_team":           resourceTeam(),
 		},
 		ConfigureFunc: configureProvider,
 	}
